refactor(components): extract text primitives (Phase 6 #2)
Splits Text, Heading, Paragraph, Code, and Divider out of builtins.tsx
into src/components/primitives/text.tsx. builtins.tsx re-exports them
so existing call sites are unchanged.

No logic changes. Refs #43.
